+++ server/src/server.ts
@@ -16,6 +16,7 @@ import * as hoverProvider from "./hover-provider";
 import * as symbolProvider from "./symbol-provider";
 import * as definitionProvider from "./definition-provider";
 import * as referencesProvider from "./references-provider";
+import * as foldingRangeProvider from "./folding-range-provider";
 import * as linter from "./linter";
 
 // Create a connection for the server (Node-IPC transport).
@@ -54,6 +55,7 @@ connection.onInitialize((params: InitializeParams) => {
       documentSymbolProvider: true,
       definitionProvider: true,
       referencesProvider: true,
+      foldingRangeProvider: true,
     },
   };
   if (hasWorkspaceFolderCapability) {
@@ -155,5 +157,7 @@ connection.onDocumentSymbol((params) => {
 });
 
 connection.onReferences(referencesProvider.resolve);
+
+connection.onFoldingRanges(foldingRangeProvider.resolve);
 
 connection.listen();